need to figure out what is wrong with Analyze negamax

+++ TakLib/IAnalysisResult.cs
@@ -2,7 +2,7 @@
 {
     public interface IAnalysisResult
     {
-        double whiteAdvantage { get; }
+        int whiteAdvantage { get; }
         GameResult gameResult { get; }
     }
 }
